+++ src/Integration/Breakdance/Compile.php
@@ -25,23 +25,29 @@ class Compile
         '_breakdance_data',
     ];
 
-    public function __invoke(): array
+    /**
+     * @param array $metadata
+     */
+    public function __invoke($metadata): array
     {
         if (! defined('__BREAKDANCE_VERSION') || (defined('BREAKDANCE_MODE') && BREAKDANCE_MODE !== 'breakdance')) {
             return [];
         }
 
-        return $this->get_contents();
+        return $this->get_contents($metadata);
     }
 
-    public function get_contents(): array
+    public function get_contents($metadata): array
     {
         $contents = [];
 
         $post_types = apply_filters('f!windpress/integration/breakdance/compile:get_contents.post_types', \Breakdance\Settings\get_allowed_post_types());
 
+        $next_batch = $metadata['next_batch'] !== false ? $metadata['next_batch'] : 1;
+
         $wpQuery = new WP_Query([
-            'posts_per_page' => -1,
+            'posts_per_page' => apply_filters('f!windpress/integration/breakdance/compile:get_contents.post_per_page', (int) get_option('posts_per_page', 20)),
+            'paged' => $next_batch,
             'fields' => 'ids',
             'post_type' => $post_types,
             // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query -- This only run by trigger on specific event
@@ -56,7 +62,13 @@ class Compile
             $contents = [...$contents, ...$this->get_post_metas($post_id)];
         }
 
-        return $contents;
+        return [
+            'metadata' => [
+                'next_batch' => $wpQuery->max_num_pages > $next_batch ? $next_batch + 1 : false,
+                'total_batches' => $wpQuery->max_num_pages,
+            ],
+            'contents' => $contents,
+        ];
     }
 
     public function get_post_metas($post_id): array