+++ src/Integration/OxygenClassic/Compile.php
@@ -34,16 +34,19 @@ class Compile
         'ct_builder_json',
     ];
 
-    public function __invoke(): array
+    /**
+     * @param array $metadata
+     */
+    public function __invoke($metadata): array
     {
         if (! defined('CT_PLUGIN_MAIN_FILE')) {
             return [];
         }
 
-        return $this->get_contents();
+        return $this->get_contents($metadata);
     }
 
-    public function get_contents(): array
+    public function get_contents($metadata): array
     {
         $contents = [];
 
@@ -53,8 +56,11 @@ class Compile
                 && get_option('oxygen_vsb_ignore_post_type_' . $post_type) !== 'true'
         );
 
+        $next_batch = $metadata['next_batch'] !== false ? $metadata['next_batch'] : 1;
+
         $wpQuery = new WP_Query([
-            'posts_per_page' => -1,
+            'posts_per_page' => apply_filters('f!windpress/integration/oxygen/compile:get_contents.post_per_page', (int) get_option('posts_per_page', 20)),
+            'paged' => $next_batch,
             'fields' => 'ids',
             'post_type' => $post_types,
             // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query -- This only run by trigger on specific event
@@ -74,7 +80,13 @@ class Compile
             $contents = [...$contents, ...$this->get_post_metas($post_id)];
         }
 
-        return $contents;
+        return [
+            'metadata' => [
+                'next_batch' => $wpQuery->max_num_pages > $next_batch ? $next_batch + 1 : false,
+                'total_batches' => $wpQuery->max_num_pages,
+            ],
+            'contents' => $contents,
+        ];
     }
 
     public function get_post_metas($post_id): array