+++ src/Integration/Timber/Compile.php
@@ -22,16 +22,19 @@ use Timber\Timber;
  */
 class Compile
 {
-    public function __invoke(): array
+    /**
+     * @param array $metadata
+     */
+    public function __invoke($metadata): array
     {
         if (! class_exists(Timber::class)) {
             return [];
         }
 
-        return $this->get_contents();
+        return $this->get_contents($metadata);
     }
 
-    public function get_contents(): array
+    public function get_contents($metadata): array
     {
         $contents = [];
 
@@ -59,6 +62,12 @@ class Compile
             ];
         }
 
-        return $contents;
+        return [
+            'metadata' => [
+                'next_batch' => false,
+                'total_batches' => 1,
+            ],
+            'contents' => $contents,
+        ];
     }
 }